rename telemetry to insights
