+++ insights/src/main/resources/META-INF/services/org.jboss.as.controller.Extension
@@ -1,0 +1,1 @@
+org.jboss.as.insights.extension.InsightsExtension
