+++ insights/src/main/resources/org/jboss/as/insights/extension/LocalDescriptions.properties
@@ -1,0 +1,13 @@
+insights=insight subsystem
+insights.add=Operation Adds subsystem
+insights.remove=Operation Removes subsystem
+insights.frequency=frequency
+insights.enabled=true or false to enable or disable the subsystem
+insights.enable=enables system and takes in the users RHN UID and RHN PW; can also take the proxyUrl and proxyPort in addition to rhnUid and rhnPw; the proxyPassword and proxyUser can be included as well.
+insights.enable.rhnUid=RHN login user ID
+insights.enable.rhnPw=RHN login password
+insights.enable.proxyPassword=proxy password
+insights.enable.proxyUser=proxy user
+insights.enable.proxyUrl=proxy URL
+insights.enable.proxyPort=proxy port
+insights.enabled=true or false to enable or disable the subsystem
