+++ insights/src/main/resources/schema/insights.xsd
@@ -1,0 +1,9 @@
+<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema"
+            targetNamespace="urn:org.jboss.as.insights:1.0"
+            xmlns="urn:org.jboss.as.insights:1.0"
+            elementFormDefault="qualified"
+            attributeFormDefault="unqualified"
+            version="1.0">
+
+    <xs:element name="insights" type="subsystemType"/>
+</xs:schema>
