+++ redhat-support-lib/src/main/java/com/redhat/gss/redhat_support_lib/api/API.java
@@ -17,7 +17,7 @@ import com.redhat.gss.redhat_support_lib.infrastructure.Search;
 import com.redhat.gss.redhat_support_lib.infrastructure.Solutions;
 import com.redhat.gss.redhat_support_lib.infrastructure.Groups;
 import com.redhat.gss.redhat_support_lib.infrastructure.Symptoms;
-import com.redhat.gss.redhat_support_lib.infrastructure.Telemetries;
+import com.redhat.gss.redhat_support_lib.infrastructure.Insights;
 import com.redhat.gss.redhat_support_lib.web.ConnectionManager;
 
 import javax.ws.rs.core.Cookie;
@@ -38,7 +38,7 @@ public class API {
     Ping ping = null;
     Groups groups = null;
     Symptoms symptoms = null;
-    Telemetries telemetries = null;
+    Insights insights = null;
 
     /**
      * @param username
@@ -76,7 +76,7 @@ public class API {
         ping = new Ping(connectionManager);
         groups = new Groups(connectionManager);
         symptoms = new Symptoms(connectionManager);
-        telemetries = new Telemetries(connectionManager);
+        insights = new Insights(connectionManager);
     }
 
     /**
@@ -118,7 +118,7 @@ public class API {
         ping = new Ping(connectionManager);
         groups = new Groups(connectionManager);
         symptoms = new Symptoms(connectionManager);
-        telemetries = new Telemetries(connectionManager);
+        insights = new Insights(connectionManager);
     }
 
     /**
@@ -163,7 +163,7 @@ public class API {
         ping = new Ping(connectionManager);
         groups = new Groups(connectionManager);
         symptoms = new Symptoms(connectionManager);
-        telemetries = new Telemetries(connectionManager);
+        insights = new Insights(connectionManager);
     }
 
     public API(String url, String proxyUser, String proxyPassword,
@@ -185,7 +185,7 @@ public class API {
         ping = new Ping(connectionManager);
         groups = new Groups(connectionManager);
         symptoms = new Symptoms(connectionManager);
-        telemetries = new Telemetries(connectionManager);
+        insights = new Insights(connectionManager);
     }
 
     public API(String username, String password, String url,
@@ -209,7 +209,7 @@ public class API {
         ping = new Ping(connectionManager);
         groups = new Groups(connectionManager);
         symptoms = new Symptoms(connectionManager);
-        telemetries = new Telemetries(connectionManager);
+        insights = new Insights(connectionManager);
     }
 
     public API(String configFileName) throws IOException {
@@ -228,7 +228,7 @@ public class API {
         ping = new Ping(connectionManager);
         groups = new Groups(connectionManager);
         symptoms = new Symptoms(connectionManager);
-        telemetries = new Telemetries(connectionManager);
+        insights = new Insights(connectionManager);
     }
 
     public Search getSearch() {
@@ -283,7 +283,7 @@ public class API {
         return symptoms;
     }
 
-    public Telemetries getTelemetries() {
-        return telemetries;
+    public Insights getInsights() {
+        return insights;
     }
 }
