--- telemetry/src/main/resources/META-INF/services/org.jboss.as.controller.Extension
@@ -1,1 +1,0 @@
-org.jboss.as.telemetry.extension.TelemetryExtension
