--- telemetry/src/main/resources/module/main/module.xml
@@ -1,19 +1,0 @@
-<module xmlns="urn:jboss:module:1.0" name="org.jboss.as.telemetry">
-    <resources>
-        <resource-root path="jboss-as-telemetry.jar"/>
-    </resources>
-    
-    <dependencies>
-        <module name="com.redhat.gss.redhat-support-lib"/>
-        <module name="javax.api"/>
-        <module name="javax.ws.rs.api" services="import"/>
-        <module name="org.jboss.staxmapper"/>
-        <module name="org.jboss.as.controller"/>
-	    <module name="org.jboss.as.jdr"/>
-        <module name="org.jboss.as.server"/>
-        <module name="org.jboss.modules"/>
-        <module name="org.jboss.msc"/>
-        <module name="org.jboss.logging"/>
-        <module name="org.jboss.vfs"/>
-    </dependencies>
-</module>
