--- telemetry/src/main/resources/org/jboss/as/telemetry/extension/LocalDescriptions.properties
@@ -1,13 +1,0 @@
-telemetry=telemetry subsystem
-telemetry.add=Operation Adds subsystem
-telemetry.remove=Operation Removes subsystem
-telemetry.frequency=frequency
-telemetry.enabled=true or false to enable or disable the subsystem
-telemetry.enable=enables system and takes in the users RHN UID and RHN PW; can also take the proxyUrl and proxyPort in addition to rhnUid and rhnPw; the proxyPassword and proxyUser can be included as well.
-telemetry.enable.rhnUid=RHN login user ID
-telemetry.enable.rhnPw=RHN login password
-telemetry.enable.proxyPassword=proxy password
-telemetry.enable.proxyUser=proxy user
-telemetry.enable.proxyUrl=proxy URL
-telemetry.enable.proxyPort=proxy port
-telemetry.enabled=true or false to enable or disable the subsystem
