--- telemetry/src/main/resources/schema/telemetry.xsd
@@ -1,9 +1,0 @@
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema"
-            targetNamespace="urn:org.jboss.as.telemetry:1.0"
-            xmlns="urn:org.jboss.as.telemetry:1.0"
-            elementFormDefault="qualified"
-            attributeFormDefault="unqualified"
-            version="1.0">
-
-    <xs:element name="telemetry" type="subsystemType"/>
-</xs:schema>
